Move main script classes into main script builder for refactoring
