--- SOC/Core/Classes/Lua/MainLuaComponents/Functions/CheckQuestAllTargetDynamic.cs
@@ -1,54 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace SOC.Classes.Lua
-{
-    static class CheckQuestAllTargetDynamic
-    {
-        public static readonly LuaFunctionOldFormat function = new LuaFunctionOldFormat("CheckQuestAllTargetDynamic", new string[] { "messageId", "gameId", "checkAnimalId" }, "" +
-            @"local currentQuestName=TppQuest.GetCurrentQuestName()
-  if TppQuest.IsEnd(currentQuestName) then
-    return TppDefine.QUEST_CLEAR_TYPE.NONE
-  end
-
-  local inTargetList = false
-  local intendedTarget = true
-  for _, CheckMethods in ipairs(CheckQuestMethodList) do
-    inTargetList, intendedTarget = CheckMethods.IsTargetSetMessageMethod(gameId, messageId, checkAnimalId)
-    if inTargetList == true then
-      break
-    end
-  end
-
-  if inTargetList == false then
-    return TppDefine.QUEST_CLEAR_TYPE.NONE
-  end
-
-  local totalTargets = 0
-  local objectiveCompleteCount = 0
-  local objectiveFailedCount = 0
-  for _, CheckMethods in ipairs(CheckQuestMethodList) do
-    totalTargets, objectiveCompleteCount, objectiveFailedCount = CheckMethods.TallyMethod(totalTargets, objectiveCompleteCount, objectiveFailedCount)
-  end
-
-  if totalTargets > 0 then
-    if objectiveCompleteCount >= totalTargets then
-      return TppDefine.QUEST_CLEAR_TYPE.CLEAR
-    elseif objectiveFailedCount > 0 then
-      return TppDefine.QUEST_CLEAR_TYPE.FAILURE
-    elseif objectiveCompleteCount > 0 then
-      if intendedTarget == true then
-        local showAnnounceLogId=TppQuest.questCompleteLangIds[TppQuest.GetCurrentQuestName()]
-        if showAnnounceLogId then
-          TppUI.ShowAnnounceLog(showAnnounceLogId,objectiveCompleteCount,totalTargets)
-        end
-      end
-    end
-  end
-  return TppDefine.QUEST_CLEAR_TYPE.NONE");
-    }
-}
