--- SOC/Core/Classes/Lua/MainLuaComponents/Functions/CheckQuestAnimal.cs
@@ -1,71 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace SOC.Classes.Lua
-{
-    class CheckQuestAnimal : CheckQuestMethodsPair
-    {
-        static readonly LuaFunctionOldFormat IsTargetSetMessageIdForAnimal = new LuaFunctionOldFormat("IsTargetSetMessageIdForAnimal", new string[] { "gameId", "messageId", "checkAnimalId" },
-    @"if checkAnimalId ~= nil then
-    local databaseId = TppAnimal.GetDataBaseIdFromAnimalId(checkAnimalId)
-    local isTarget = false
-    for animalId, targetInfo in pairs(mvars.ani_questTargetList) do
-      if targetInfo.idType == ""animalId"" then
-        if animalId == checkAnimalId then
-          targetInfo.messageId = messageId or ""None""
-            isTarget = true
-          end
-        elseif targetInfo.idType == ""databaseId"" then
-          if animalId == databaseId then
-            targetInfo.messageId = messageId or ""None""
-            isTarget = true
-          end
-        elseif targetInfo.idType == ""targetName"" then
-          local animalGameId = GetGameObjectId(animalId)
-          if animalGameId == gameId then
-            targetInfo.messageId = messageId
-            isTarget = true
-          end
-        end
-      end
-      return isTarget, true
-    end
-  return false, false");
-
-        static readonly LuaFunctionOldFormat TallyAnimalTargets = new LuaFunctionOldFormat("TallyAnimalTargets", new string[] { "totalTargets", "objectiveCompleteCount", "objectiveFailedCount" },
-            @" local dynamicQuestType = ObjectiveTypeList.animalObjective
-  for animalId, targetInfo in pairs(mvars.ani_questTargetList) do
-    local targetMessageId = targetInfo.messageId
-
-    if targetMessageId ~= ""None"" then
-      if dynamicQuestType == RECOVERED then
-        if (targetMessageId == ""Fulton"") then
-          objectiveCompleteCount = objectiveCompleteCount + 1
-        elseif (targetMessageId == ""FultonFailed"") or (targetMessageId == ""Dead"") then
-          objectiveFailedCount = objectiveFailedCount + 1
-        end
-
-      elseif dynamicQuestType == ELIMINATE then
-        if (targetMessageId == ""Fulton"") or (targetMessageId == ""FultonFailed"") or (targetMessageId == ""Dead"") then
-          objectiveCompleteCount = objectiveCompleteCount + 1
-        end
-
-      elseif dynamicQuestType == KILLREQUIRED then
-        if (targetMessageId == ""FultonFailed"") or (targetMessageId == ""Dead"") then
-          objectiveCompleteCount = objectiveCompleteCount + 1
-        elseif (targetMessageId == ""Fulton"") then
-          objectiveFailedCount = objectiveFailedCount + 1
-        end
-      end
-    end
-    totalTargets = totalTargets + 1
-  end
-  return totalTargets, objectiveCompleteCount, objectiveFailedCount; ");
-
-        public CheckQuestAnimal(MainScriptBuilder mainLua, string objectiveType) : base(mainLua, IsTargetSetMessageIdForAnimal, TallyAnimalTargets, "animalObjective = " + objectiveType) { }
-    }
-}
