--- SOC/Core/Classes/Lua/MainLuaComponents/Functions/Messages.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOC.Classes.Lua
-{
-    class Messages
-    {
-        public string ToLua(MainScriptBuilder mainLua)
-        {
-            return @"
-this.Messages = function()
-  return
-    StrCode32Table {
-      Block = {
-        {
-          msg = ""StageBlockCurrentSmallBlockIndexUpdated"",
-          func = function() end,
-        },
-      },
-    }
-end";
-        }
-    }
-}
