--- SOC/Core/Classes/Lua/MainLuaComponents/Functions/OnTerminate.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SOC.Classes.Lua
-{
-    class OnTerminate
-    {
-        public string ToLua(MainScriptBuilder mainLua)
-        {
-            return @"
-function this.OnTerminate()
-	TppQuest.QuestBlockOnTerminate(this)
-end";
-        }
-    }
-}
