+++ SOC/SOC.csproj
@@ -139,28 +139,9 @@
     <Compile Include="Core\Classes\Lua\LuaTemplate.cs" />
     <Compile Include="Core\Classes\Lua\LuaText.cs" />
     <Compile Include="Core\Classes\QuestBuild\Builders\DefinitionScriptBuilder.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\AuxiliaryCode.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\CheckQuestAllTargetDynamic.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\CheckQuestAnimal.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\CheckQuestGenericEnemy.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\CheckQuestItem.cs" />
     <Compile Include="Core\Classes\Lua\LuaValue.cs" />
     <Compile Include="Core\Classes\Lua\LuaVariable.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\QStep_MainCommonMessages.cs" />
-    <Compile Include="Core\Classes\Lua\LuaFunctionOldFormat.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\CheckQuestMethodsList.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\Messages.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\OpeningVariables.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\GenericTargetPair.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\ObjectiveTypesList.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\OnAllocate.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\OnInitialize.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\OnTerminate.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\OnUpdate.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\QStep_Main.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\Functions\QStep_Start.cs" />
     <Compile Include="Core\Classes\QuestBuild\Builders\MainScriptBuilder.cs" />
-    <Compile Include="Core\Classes\Lua\MainLuaComponents\QuestTable.cs" />
     <Compile Include="Core\Classes\QuestBuild\Builders\CommonAssetsBuilder.cs" />
     <Compile Include="Core\Classes\QuestBuild\BuildManager.cs" />
     <Compile Include="Core\Classes\Common\LoadAreas.cs" />
